- Work on parser & project model classes + integration with data storage classes (for from done but good progress) - a bit more work on the ProjectController - added documentation folder - added XML format to documentation

+++ ExCiteS Data Storage/src/uk/ac/ucl/excites/storage/model/Record.java
@@ -23,7 +23,7 @@ public class Record
 		this(schema, UNKNOWN_ID, UNKNOWN_ID);
 	}
 	
-	public Record(Schema schema, int deviceID)
+	public Record(Schema schema, long deviceID)
 	{
 		this(schema, UNKNOWN_ID, deviceID);
 	}
